Nullish NBTData Properties
Properly moving over `NBTData`'s implementation of format property descriptions so they are consistently present on the object itself. Right now, for 'optional' format types like Compression, and Bedrock Level, they will not be added to the object if they don't have a value. I'm going to change it over so that these will instead be set to `null` if they don't have a specific value applied to them by the file, or the user that constructed the object. I found out a nice way to configure the properties' read, write, and enumerability access, and this essentially makes it so the user can now extend the object as they may like to. Now the existing properties on the object are only made read-only, and you are free to do anything else on the object. I was using `Object.freeze()` before, which would break extensibility.

Alongside this, I want to better allow the use of schemas for the different file formats, so you'd be able to use a single configuration object to read and write binary NBT, and create your own `NBTData` objects. So I want to update `ReadOptions` so it can accept the same shape as an `NBTData` object, which would then allow you to have a single configuration shape for all API endpoints!