Function Options Overloads
I like how this looks for `NBTData`, but I'm not sure if it's worth it to do for other regular functions too. You can hide the destructured function parameters from the type definition, and simply show it as a plain object type instead. This is neater in terms of how the documentation looks, but I don't like that it makes the library code more obtuse. Another option I could do is move the destructuring to inside the function body itself, which will hide the more in-depth behaviors from the documentation. I think this is more of a concern for me now because the types for things are getting a lot longer now that I'm adding generics in there, and I don't want things to be too long to be able to read with ease in the editor.

https://www.typescriptlang.org/docs/handbook/2/conditional-types.html#conditional-type-constraints
https://www.typescriptlang.org/docs/handbook/2/generics.html#generic-classes
https://www.typescriptlang.org/docs/handbook/2/generics.html#using-type-parameters-in-generic-constraints
https://blog.logrocket.com/understanding-infer-typescript/

I wanted to try flattening the data parameter for the `NBTData` constructor, using overloads, but that didn't seem to work when using `NBTData<T>` on it's own. Same with using `NBTData<infer U>`, looks like the class generic type parameters only work if it's the raw type parameter itself, it can't be wrapped by another type? I just don't like that it shows the value of the `NBTData` data three times in the IDE hinting; once for the generic, another for the type parameter, and the third for the `NBTData<T>` portion. So I wanted to essentially move `NBTData<T>` to it's own overload, but that prevented the `T` unwrapping from working properly, meaning that it resulted in a double-wrapped `NBTData`, `NBTData<NBTData<NBTShape>>`.