Deflate-Raw Support
Added native support for Deflate-Raw into the library itself! Now it's an actual compression type that can be used with all of the other APIs that specify a compression type.

I also brought back the Compression module to now be a more combined version that's right in the middle of the original logic/APIs, and with the new performant stream chunking implementation that I added in the last few updates. I'm deciding to still keep NBTify-only compression functions within the library, as it isn't too far fetched to have those nicely embedded into the library, especially for once the Compression Streams API support status is more streamlined across the different JS platforms and versions. Right now the only hangups in terms of support are Node's missing `deflate-raw` support, and the older versions of browsers which didn't have support for the APIs yet. All of the modern browser vendors support them now, but only as of the last release or two, and that's not quite up to my compatibility standards, since that is a smaller part of the entire user bases for those browsers. I want the majority of users to be able to use whatever the thing I'm making is. So, if that involves adding in a polyfill, then that's ok. I'd just rather it successfully work for the majority, instead of work without a polyfill for the minority.