Improved Unsafe Typings
Realized how I can keep kind of enable user-facing type checking for `CompoundTag`s / `RootTag`s. It doesn't work at the parameter level yet, as that's not something TypeScript can currently do unfortunately, but you can add a `satisfies` operator after the value you'd like to verify, which is a nice fallback!

Moved back the Tag module to how I used to structure it. I think I thought I didn't like how I had it organized before, but I think it probably was as nice as it had to be.

Listening to Wine and Pickles.