Node Inspect Lib Types
After building nad publishing the previous update, I discovered that the public-facing types for the library depended on having the types for Node installed alongside with the library itself, which I want to try and omit. I don't want extra library types for specific features, I'd rather those only be used in the dev environment to build the library itself, they shouldn't be required to use the library. This makes it so that the types can't be included in the final project build, and it also ensures that the types that are used for function return types are hidden from the public build as well.

I don't like that the private methods are marked as unused, but I haven't found a different way around that quite yet. Naming the symbol with an underscore doesn't appear to work, and adding an overload to the getter with an `unknown` return type doesn't work either, as you can't add overloads to getters it seems like. I'm glad that this fixes things at least! If this didn't work, I could also probably just assign the getter to the prototype using the constructor instead of on the class (this will skip TypeScript's checking, since the use of `Object.defineProperty()`), or I could just add `@types/node` as a regular dependency instead of only for `dev`. That's the least ideal solution, but it is one to keep in mind either way. I've learned to remember that more often with random things, to consider all options even if some of them are 'obviously' not the right choice. It's still an option, you don't know what you may think of while considering all of them!

Also, to note: About the articles below, I combined two methods to fix this. First was the dynamic type import, which hides the dependency from being visible as a package in implementer libraries (doesn't solve the output types references though, it only makes it so it doesn't 'require' the types). The full fix was the marking the methods as private, since that will disable the return types from being public on the output, hence making the references not broken, because the Node-only types aren't referenced anywhere publicly in the library's types.

https://davidea.st/articles/typescript-2-9-import-types/
https://stackoverflow.com/questions/69666568/how-can-i-use-the-types-of-a-conditionally-imported-module-in-typescript