Ridiculous Fix
A clever name, haha. While working on bringing these updates over to Dovetail, I discovered a bug with opening unnamed root tag files. Turns out when working on the code for the `SIngle-Definition Configuration Schema`, I accidentally changed the unnamed root tag auto-check step, meaning the `read()` function was only able to open named root files. This showed up as an issue while opening `ridiculous.nbt` into Dovetail, to which I was confused as to how it was having trouble opening the file.

I think this is an example of why I need to implement tests for the library, at least at the file format level. I can have a simple file type for most kinds of formats, and I can check that the reader and writer correctly open and save to and from each format as expected. This issue is a big one, and I essentially only found it in production, or, almost-production. I think it also points out that my auto-detection code could be a little less verbose too, as it kind of hid in the rest of the code there. Or, things like this just tend to happen. I can't plan for every issue I make, so maybe I just have to try my best and deal with it, and double check everything I'm doing. At the same time, I think I can do more than just that. I think a combination of that acceptance, along with some simple tests could be a strong way to go :)

Listening to Wine and Pickles again, 'Inhale'. Now '4s' is starting, and it makes me want to somehow show Elkosaurus Tex to Mike Keneally, I think he'd get a kick out of that project of mine!