Return Types, API
Improved the library's IO by adding more explicit return types for functions and the like. I also discovered that I forgot to add the `NBTData`-like equivalent generics for SNBT parsing. Since that returns the plain `CompoundTag` data itself (not wrapped in an `NBTData` object), the generic is simply for that, rather than for the internal shape of an `NBTData` object, since that's not what it returns.

Found a few internal spots within the library that could manage the types for tags a bit better. Essentially, I widened a few of the functions so that they can use NBTify's tag primitives in a more streamlined manner, to where it doesn't require the libraries uses for those functions to shape the data to go into the function, instead the function does that! Essentially, things like the NBTify primitive wrapper classes (`Int8`, `Int32`, ...) are now accepted into the `NBTWriter` class methods, rather than requiring `#writeTag()` to handle the formatting.

Also a nice thing with doing this, it means that now the library can check itself against the types for each tag more thoroughly, since they specify where each thing should be able to handle each tag. This is part of why I wanted to define types for each of the JavaScript-side of NBT tags, because it allows me to add or change what each tag is built with in the future, and including that the types will show errors where the library's implementations can't handle that new primitive addition. In a smaller way, if I wanted to add a new class type to the `ByteTag` type, I could do that and any code in the library that doesn't yet support that new additional type inside of the union, then the internal handling for that tag will show where that new value kind won't be anticipated correctly. Then I'd just have to add the new handling where it flags it as an issue.

This works the same way around, for the user too. If I changed the backing of an NBTify primitive to a different representation, then the type definitions for any given NBT structure can stay safe, because they are built with whatever aliases NBTify uses (If you build your NBT structure types using NBTify primitives, they don't have to know what NBTify actually uses to manage NBT data in the JavaScript-land), so then the user's code would also show errors if I changed a primitive respresentation out to something else.

Ok, a real example after that jargon. If JavaScript were to add a new `int8` type, that's in similar behavior to the current `number` type, I could remove NBTify's custom `Int8` class, and instead use the new built-in `int8` number type instead (Or, make `ByteTag` a union type which accepts either `int8` or `Int8`, that's the more realistic behavior to help users).

Had NBTify been something I made a while back, this could have been the same case as before `bigint` was introduced. I likely would have needed to use something like the Long library, to represent the `LongTag` type. In what would have been the migration to `bigint`, I could have done the same thing. This also works for the TypedArray-based tag primitives, for from before TypedArrays were added.

I was planning on working on the new SNBT parsing originally, but I started seeing where these changes could be of help in getting things set up nice first. Strengthening the library's use of it's own primitive types seemed like a strong move to make sure any changes I make are correctly reflected throughout the codebase.