SNBT Testing Demo
Going to start working on an updated SNBT parser next, this is the config I set up to try it out.

Realized the only reason files compressed with zlib in Node have different content across platforms (Linux, and macOS at least) is that it's the compression header that only changes. If you mask out the header for the files in part of the comparison call, then that's the only part that changes! Really neat find. So this is a more accurate way to check the files are the same, as it won't throw false *negatives*, because they are actually the same in most cases, the only change is the file header because you are writing with a different zlib implementation than to that of what the original file was written on (I think?)