SNBT Writing Strength
Buffing up the current implementation's strength before I try to do my own from scratch. I'm working on learning more about how this setup works before doing my own, because it tends to be that more things get missed if you don't have as much of an understanding about what you're trying to build as what you may need to have for it to work just right.

Looking into what the different options are for the lexing process, as I'm not sure whether I should be using strings, or TypedArrays. Either way, I have to read it as text, but I'm wondering which is a better idea to read from as a whole? Probably just strings I'd imagine?

https://stackoverflow.com/questions/45803829/memory-overhead-of-typed-arrays-vs-strings
https://areknawo.com/the-lexer-in-js/
https://codemaster138.github.io/blog/creating-an-interpreter-part-1-lexer/

As another reference to it again, this is one of the main videos I've been watching to figure out how I should build my lexer. In the beginning of the video, he mentions that doing a split on the string isn't likely the most performant option to do, and that there's probably a more performant way, and that he did the easier way for the video. I'm curious what way he's talking about is, because most references online say that using string splitting is actually really performant, so I don't know for sure which way to go yet. Maybe to just iterate over the one string as a whole instead of splitting it into it's individual characters?

https://www.youtube.com/watch?v=8VB5TY1sIRo