Second Stage: Initial Read Testing
Trying out some things here, not sure if it's the way to go. I want the SNBT parsing structures to look and feel like the rest of NBTify, in terms of what each function and method handles, and to where the dependency injection bounderies lie.

This was also from the other day, this is where I started to feel like these were changes that were building off of the previous commit.