nbt-ts SNBT Test
Made a demo that uses nbt-ts' SNBT implementation to see if it can work closely enough with NBTify's library primitives. Looks like it does! The only things are that it doesn't support the same stringifying that I added to my own implementation, and I think it doesn't quite follow the SNBT spec quite exactly like Deepslate does, notably for the TypedArray-based tags.