SNBT Test Refactor
Moved the stringify logic into it's actual module, out of the parse module. Added general code formatting to make all of the new code more structurally similar to how I like things. Moved the parse logic over to a class-based implementation, with the same internal functions now being methods on the builder class, instead of private functions within the scope of the main `parse` function.

Need to re-add the unsafe List and Compound tag handling to the stringify logic, and I discovered that NBTify's primitive number types aren't being used, and instead are being asymmetrically returned (in terms of the API representations at the type/primitive level) as a simple `string` type, which loses the specificity of the number primitive that it's representing.