Original Stringify Parsing Support
Going back though my original stringify implementation, it's actually a fair bit closer to what I want it to be, so I think it doesn't need as much of a rework as I thought it did.

With bringing that over, this reestablishes the full SNBT spec support, meaning I had to figure out how to add the missing TypedArray tag entry suffix reading checks, which appear to be working out now!

Made a few other nice typing changes to the stringifying script, mostly to do with return types and things like that. I also merged `writeBoolean` and `writeByte` into a single method.