Standalone Tag Sanitizing
The codebase has multiple places where it needs to sanitize the values in dictionary-like tag types it works with, so I thought I'd abstract this away into a single implementation, since they all do the same thing.

I do like that it puts the same logic all in one place to manage, but I don't like the abstraction effects that it puts on the rest of the places. May look into a new way to sanitize, or how to manage this handling in a better fashion.

I also updated the return types for methods in the definition generator module.