Method Primitive Consistency
This is an alternative that makes the parameter types more consistent in regards to accepting only the NBTify primitives, meaning you have to wrap the base `number` type in one of the NBTify primitives so it can be consistently represented across the codebase. Once again, I like the idea of this, but in practice, it doesn't make sense to wrap a value in a wrapper only to make the types consistent. This doesn't add any value, and it creates unnecessary objects which are just garbage collected after you unwrap them again. So I think for the custom NBTify primitive wrappers, it's ok to accept bare `number` values when they are coming directly from the library, since you already know their type at that point. It's just a matter of how they are packaged.

For both of these commits, I think documenting this is a good step, and I will find a spot in the middleground for the long term solution. Making things more complex only to make the API look nicer isn't the goal, as I instead want usability and simplicity to come first.