Lighter Public API
Scoped multiple parts of the library's implementation details to be hidden from the user-facing exports of the library. I haven't thought of a use where I or any of NBTify's users would be able to do anything with the bonus exports from the library's implementation. I'd really be interested in hearing if there is a use case for this, but so far it seems like maintaining a smaller public API seems like a nicer way to go.

Not sure how I want the module scoping to work completely, I think that's part of the challenge with hiding some of the internal usage-specific exports, as they are needed by other modules within the library. I am going to explore different ways of organizing everything, as currently I don't like how a few of the modules are structured in terms of the scope of things they define, one example being the data module, since some of it consists of types used across the library, and also the NBTData class. While they are related, I'm not sure they make sense to be together. Same think with the Tag type definitions, and the tag utility functions.

I'd rather the modules themselves be what dictate what things are part of the public API, I don't want to manually hide exports from a given module because some of it's items are publicly available, and some are library-only. Instead, those should be in their own module, which have exports available to anywhere in the library, and that module is simply excluded from `./src/index.ts`.

Some implementation are unique to individual modules, and that's ok. That's not an issue, that works great with modules! It's when more than one module in the library needs it, but I don't want it to be publicly available from the library's API.