Module Organization Experimenting
Trying to come up with how the public-facing module appearance, API structure, and types should be provided for the users of the library. Realized that the definition generator could make sense to be in the stringify module instead of by itself.

Going to try out overload signatures again, now that there's a slimmer public API surface. It's not as much to make the public types a bit nicer, since I'm using immediately-extracted object parameters. Listing all of the parameters in the type is a bit ugly, so this helps make the type look better, while still being able to have immediately destructured options object parameters. Recently adding full return types for everything helped making this more streamlined too.

Also realized that I could abstract part of the reading process' primitive wrapping into the reading methods themselves. While I don't like the overload parts, it does make it specific as to what it's doing in terms of the type level.