Standard Lib: Compression Typings!
TypeScript added type definitions for the Compression Streams API, yay!!!

As of when committing this, VSCode's TypeScript version doesn't updated to support this yet, so you may have to specify the editor lib version using a config file, which can be set using the Command Palette to change the TypeScript version to use that of the workspace. Otherwise, placing this in `./.vscode/settings.json` will do the same thing.

```json
{
  "typescript.tsdk": "node_modules/typescript/lib"
}
```

https://github.com/Offroaders123/Dovetail/issues/1