Iterator Typings
Adjusted the typings for the compression module's typings. Discovered that TSC's new Compression Streams types aren't as specific as to that of what I've been using already, and it started to leak `any`s into the codebase. Naah! Might make a PR to adjust the standard lib typings to reflect these more-specific property values.

Added a fallback to allow for the full async iterator syntax with the vanilla `ReadableStream` object, if it is supported in the current runtime. It's working in NodeJS (only) at the moment, so I thought it may be worthwhile to only polyfill the support if it isn't supported, since it partially is/will be supported in other implementations down the line.