Generic Shadows: Success!
Wow haha, turns out it was simply solveable with a slight type assertion. Didn't realize that would help there, guess I hadn't tried that yet. I think it was this error that helped me figure it out:

```
Argument of type '{ readonly name: Name; readonly endian: Endian; }' is not assignable to parameter of type 'U'.
'{ readonly name: Name; readonly endian: Endian; }' is assignable to the constraint of type 'U', but 'U' could be instantiated with a different subtype of constraint 'FormatOptions'. ts(2345)
```

I think this was the original error that I was encountering earlier. It didn't quite click for me what it was trying to say.

```
Type '{}' is not assignable to type 'U'. '{}' is assignable to the constraint of type 'U', but 'U' could be instantiated with a different subtype of constraint 'FormatOptions'. ts(2322)
```

I think I realized with the commit above that the type you are passing in is more specific than what the generic allows for, and that could break the types for the generic if you were to try narrowing the type like that. So it's not the parameter's default's issue, it's that the default is too specific for what the generic could allow for. So broadening the scope of the default back up to the type for the generic will make things dandy again. I think my initial problem was that I didn't know that it narrows the scope like that. I guess I did know that, since that's how my NBT primitive wrapper classes work with values, but I hadn't tied the scope-ish thing together like that yet.

So for the `2322` error message, I thought that assigning the default value to the parameter/variable would just use the type for that variable, but looks like there's room for two types in this case, one for what the generic is constrained to, and what the actual default is by itself. I'm curious what can be done to use this to one's *advantage?* This is making some other subconsious lightbulbs go off in my head.

I'm also wondering if there's some way I can add `const T` support to my NBT primitive wrappers, since you don't always want to have the constant value of the generic type. Say you pass in `35` as the number value, maybe you don't want to allow only `35` for that variable's value type. Right now you have to do `35 as number` or `as PrimitiveWrapper<number>` to widen the type scope back up again.

Thinking about it, it seems like these two behaviours (Why do I do that by default now, haha? Guess I'm turning British) are two sides to the same coin (Glass Attack!). I like the idea of what I could possibly use these to build things with. Super neat! Generics are complicated, and ***sick***! ('sick' was in Muscle-Man's voice, not sure why either)

Oh yeah, was working on a standalone demo to try and debug things in an effort to getting it to work. I'm glad I tried again, I was almost going to move on instead. I think using this in *combination* with the function type handling, this will be even stronger to build things with!

I'm glad that I dove this deep into getting the types for things set up, and that TypeScript can allow for things as complicated as this. It's super nice.

```ts
export interface Formatze {
  fancy?: boolean;
  woaheiy?: string;
}

export class NoiceData<T extends object = any, const U extends Formatze = Formatze> {
  declare readonly data: T;
  declare readonly fancy: U["fancy"];
  declare readonly woaheiy: U["woaheiy"];

  constructor(data: T | NoiceData<T>, options?: U);
  constructor(data: T | NoiceData<T>, { fancy, woaheiy }: U = {} as U) {
    if (data instanceof NoiceData){
      data = data.data;
    }
    this.data = data;
    this.fancy = fancy;
    this.woaheiy = woaheiy;
  }
}

declare const noice: NoiceData;

const heya = new NoiceData({},{ fancy: true, woaheiy: "gg" });
heya.fancy;
heya.woaheiy;

// I'd like to be able to merge the properties, at the type-level too!
const oaugh = new NoiceData(heya,{  });
```