BooleanTag Specifics + OtherDocs
Was going to try adding a generic parameter to the `BooleanTag` type, but that didn't work too well. Realized that it would actually work nicely to have distinguished types between what would be considered true and false independently, and you can instead use that to declare if a `BooleanTag` is only true or false. I haven't had an occurrence of this being needed for any property or tag definitions yet, but it doesn't mean it won't be needed. I liked the idea of using a generic to force the boolean value, but it didn't seem to work like I thought it would.

After typing that last part, I got a working solution for it, this time a little bit different:

```ts
export type BooleanTag<T extends boolean = boolean> = T extends false ? false | ByteTag<0> : T extends true ? true | ByteTag<1> : boolean | ByteTag<0> | ByteTag<1>;
```

It does work, but I don't like the verbosity, and it isn't as explicit as just changing the 'tag' to the other state. So, might as well go with the same thing most of the rest of NBTify follows, as using the 'thing' to describe the type, rather than a property on the thing itself.

Found some other documentation I forgot to re-add when doing the SNBT parser rework.

Also found out why the `NBTData` constructor generic wasn't working before, and realized that leaving the generic parameter as-is is fine when you couple that *with* the options object type assertion. Being more explicit, for the win! This is the f****** way to go!! Haha, more explicit for the win. (That's actually bleeped out, it's not GitHub lol)