Format Generic Merging Issues
Two things: Tried setting the default types for the `NBTData` constructor generic to what the `NBTData` object applies by default, but that narrows the type autocompletion down to that of the default, which isn't ideal. So it would default to `{ name: ""; endian: "big"; compression: null; bedrockLevel: null; }`, but that narrowed down the option values to those only, rather than the union types that each property provides. It only gave `"big"` endian as an option, for example.

The other one I discovered the other day, that doesn't work here either, is that it currently doesn't know how to merge the inferences together. I'd like the manually provided properties to use the values that you gave to the constructor/generic, but the non-provided ones should fall back to the union types as the default. For example, only passing `{ endian: "little"; }`, the `compression` property should still be of type `Compression`, while it actually goes to `unknown` with this current implementation. I tried doing an `U["-property-"] extends unknown ? PropertyType : U["-property-"]`, but I think that forced it to be `unknown` all of the time. Gonna set that aside for now, and see what happens when I look at it a few more times.

I think this is kind of what I'm dealing with, but I'm not totally sure.

https://stackoverflow.com/questions/73215573/how-to-return-type-based-on-the-object-properties-parameter-with-generic-type