Format Generic Merging! (Very Close)
Searched a bit for this kind of setup, and I found a demo on StackOverflow that did exactly what I wanted to have!

It very closely does everything I need it to. It properly merges the options types between the options objects. The only missing inference is when you pass in an `NBTData` object into the options object, rather than a plain options object. I haven't figured out/discovered how to infer the type from THAT one's options generic, so it only shows the types defined from that second constructor. This is very close though! I'm happy this is working.

The other thing is that requires you to specify the return type on the class's constructor, which you unfortunately can't do with the ES6 class/constructor syntax. So I had to define the `NBTData` object using interface types, which will allow you to specify the constructor's return type. I think this is necessary to narrow the type when you construct the class, as I'm not sure this is possible to do with the ES6 syntax in terms of generics' handling.

There were LOTS of resources for this one. I've been watching videos the last few nights, and I've wanted to add them to my knowledge catalog here, but I haven't had any commits since seeing them.

For this feature:
https://stackoverflow.com/questions/49682569/typescript-merge-object-types (THANK YOU!!)
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-1.html#recursive-conditional-types

Cool finds:
https://www.youtube.com/watch?v=MToTEqoVv3I (This was a cool video and it felt really relatable haha)
https://www.youtube.com/watch?v=X6fF0QXUACQ (Same here! Found my 'people' haha)
https://www.youtube.com/watch?v=LJds_OzSag8 (yes)
https://www.youtube.com/watch?v=rz0cY7AYZNQ (yes,yes)
https://www.youtube.com/watch?v=gUGXwb0qEAg (another cool crossover, didn't know he likes Steven's work! I've watched him a few times over the years)
https://www.youtube.com/watch?v=Tml94je2edk (an interesting find)

This has to do with issue #29, so I'll mention that here too.

About that last link, I think TypeScript is in that sweet spot right in the middle. The dynamic-nature of JavaScript is appealing because you don't need the type to *run* the program. If it did have types in the syntax, it would be perfect for linting and type 'checking', as you are essentially having the runtime delegate the type handling, rather than your documentation for you variables. Mentioning it like that, it sounds less efficient (well, I guess it is, but it depends whether you're talking about run-time speed, or DX).

I think this is a really nice trade-off though, at least in that just like mentioned in the video, you get the best of both worlds. Your types don't have to be correct to run the program, they are only to check that things are 'working' right. In fully static-typed languages, the types are the things that dictate how the program runs, it's both for the run-time implications, and for the code documentation.

So I think the argument is less about documenting your code with types, and more about whether the code should be dynamically compiled, or uses the types to generate the runnable executable output. I think this is where the actual trade-off is.

I think having types to define your code structures has shown (for me at least, and probably by others) to be universally better to help maintain the quality of your code, as you can keep track of things much better. If I were to (or probably, once I eventually try to) create my own language/syntax, I think having statically-typed code, which runs on a dynamic runtime is the way I'd go, in the event that max performance isn't the main goal.

I'm not saying that full performance isn't as important out of these trade-offs, just that saying 'types in syntax' vs 'no types in syntax' isn't really the issue, it's moreso the compilation/runtime behavior which declares that, not the code syntax. Of course they implicitly line up because of some things (I don't think you could have statically-typed dynamic code, if that makes sense. I'm thinking like a table of 'yes' or 'no' things about either given ideology. Dynamic code doesn't require 'yes' types, or 'no' types. Statically-typed code must require types, only being 'yes'. You can't have a statically-typed language without explicit return types, hence making 'no' impossible. This sounds like a simple example of an 'additive' theory or something. Not sure if that's even anything, but that's what I mean).