Format Options: Fresh Start + New Ideas
Realized a possibly better way to implement this kind of thing, and it involves abstracting away the use of the `NBTData` constructor itself. Instead, NBTify will provide a top-level export for a function named `build()`, instead (tentative naming). This will simply construct an `NBTData` object for you, rather than requiring the use of the `new` keyword (not a goal, but a nice side-effect too). This will fix the problem I encountered in the previous four commits, having to do with the explicit return type generics for the `NBTData` constructor, which can't be asserted at the type-level unfortunately, because you can't add a return type annotation to a class constructor. I am instead able to do this if I am to wrap the use of the `NBTData` constructor in a regular function, which will allow for a custom return type annotation. This is necessary for the options object merging because you have to perform a type merging of the two generic types, and this doesn't appear to be doable directly in a class definition's generic syntax section.

I also want to rework the use of the `NBTData` class though too, as some of the functionality that I want to ensure both at the type level and run-time level, aren't really necessary, and they are just adding extra overhead to things that don't really help with anything.

I thought it might be useful to be able to nest `NBTData` declarations, in the way of borrowing the config from one `NBTData` object, then passing it as a config to construct another, say if you wanted to borrow the same format from that one. I think this is probably less necessary of something to really be helpful, so I'm either going to remove it, or I'm going to work this behavior in as an overload to the `build()` function instead, maybe. In the previous commit, I was also going back to experiment with adding a dedicated `getFormatOptions()` method on `NBTData` object instances, which could be used independently to get the same format between two `NBTData` objects.

I think the hard part is trying to figure out how I'd want to use my API, and how others will also.

Oh, and the explicit return type annotations are actually required for the `WrapOptions`-like behavior, which will prevent the `NBTData` properties from typing-out as `unknown`, say if you only specified a single options parameter value. This was the main problem with the current generic handling setup present in the stable NBTify releases.